requirements.txt
numpy
pandas
boto3
pillow
matplotlib
mysql-connector-python-rf